+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+* v0.7.16
+- Fixed bug with `npm ERR! enoent ENOENT: no such file or directory,`
+
 * v0.7.13
 - Full typescript support! Files are Transpiled using ParcelJS insternally inside `@gapi/cli`
 - 500 Migrations run with Typescript > Javascript > Execution in 13 seconds!